Disable Touchscreen Pan-X / Pan-Y
Dragging the slider on touchscreen devices makes the browser think that the user is trying to pan/scroll the page by holding on to it. Touch events have been disabled for the slider. So that only the thumb moves rather than make it move along with the whole webpage on touchscreen devices.